+++ cuda/C_GCN_MP.h
@@ -2,7 +2,7 @@
 
 namespace CU_MP {
 
-void GCNLayer(float* edgeIndex, float* featureTensor, float *aggregationVar, float *nodeDegrees,
+void GCNLayer(int* edgeIndex, float* featureTensor, float *aggregationVar, float *nodeDegrees,
 		int numOfNodes, int numOfFeatures, int numOfEdges, int outputSize);
 __global__ void GCNLayerNew(float* edgeIndex, float* featureTensor, float *aggregationVar, float *nodeDegrees, int numOfNodes, int numOfFeatures, int numOfEdges);
